added tabs folder and moved files and changed layoutfile
