+++ frontend/app/(tabs)/NavBar.tsx
