+++ frontend/app/_layout.tsx
@@ -1,5 +1,10 @@
-import { Stack } from "expo-router";
+import { Tabs } from "expo-router";
 
 export default function RootLayout() {
-  return <Stack />;
+  return (
+    <Tabs>
+      <Tabs.Screen name="HomeTasksScreen" options={{ title: "Homepage" }} />
+      <Tabs.Screen name="NavBar" options={{ title: "NavBar" }} />
+    </Tabs>
+  );
 }
