+++ frontend/app/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import HomeTasksScreen from "../components/HomeTasksScreen";
+import HomeTasksScreen from "../app/(tabs)/HomeTasksScreen";
 
 export default function Index() {
   return <HomeTasksScreen />;
